backend_py/requirements.txt
flask
python-dotenv
langchain
langgraph
langchain-google-genai
langchain-core
langchain-text-splitters
langchain-chroma
langchain-community
gunicorn
flask-cors
elevenlabs